[netchan] Correct type of last_received
Storing a double in a float does not go well when that double starts at
4294967296. Fixes the delayed timeouts in qtv.

+++ include/netchan.h
@@ -229,7 +229,7 @@ typedef enum {
 typedef struct netchan_s {
 	qboolean	fatal_error;	///< True if the message overflowed
 
-	float		last_received;	///< Time the last packet was received.
+	double      last_received;	///< Time the last packet was received.
 
 	/// \name statistics
 	/// the statistics are cleared at each client begin, because
